[Server] Remove our custom logger
This was always a terrible hack and was causing weird things to
happen to any processes that were attempting to read from the log
file while the logger was being replaced. The main reason for adding
the logger was to have coloured console output, which is now handled
in the frontend via the new control panel.